split up config to separate files
